+++ src/main/java/coz/weavon/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package coz.weavon.config;
 
-import coz.weavon.common.application.service.MessageTranslator;
+import coz.weavon.helper.MessageTranslator;
 import coz.weavon.core.auth.application.service.AuthUserService;
 import coz.weavon.core.auth.presentation.filter.JwtAuthenticationFilter;
 import coz.weavon.core.auth.presentation.filter.UsernameAuthenticationFilter;
