+++ src/main/java/coz/weavon/core/auth/application/service/AuthUserService.java
@@ -1,6 +1,6 @@
 package coz.weavon.core.auth.application.service;
 
-import coz.weavon.common.application.service.MessageTranslator;
+import coz.weavon.helper.MessageTranslator;
 import coz.weavon.core.auth.application.adapter.AuthUserAdapter;
 import coz.weavon.core.auth.domain.model.AuthUser;
 import coz.weavon.core.auth.domain.model.OAuthUser;
