+++ src/main/java/coz/weavon/core/auth/presentation/controller/AuthController.java
@@ -1,6 +1,6 @@
 package coz.weavon.core.auth.presentation.controller;
 
-import coz.weavon.common.application.service.MessageTranslator;
+import coz.weavon.helper.MessageTranslator;
 import coz.weavon.common.presentation.model.response.RestResponse;
 import coz.weavon.core.auth.application.service.AuthService;
 import coz.weavon.core.auth.domain.model.AuthUser;
