+++ src/main/java/coz/weavon/exception/handler/GlobalExceptionHandler.java
@@ -4,7 +4,7 @@ import coz.weavon.exception.model.AuthException;
 import coz.weavon.exception.model.BusinessException;
 import coz.weavon.exception.model.ClientException;
 import coz.weavon.exception.model.RestException;
-import coz.weavon.common.application.service.MessageTranslator;
+import coz.weavon.helper.MessageTranslator;
 import coz.weavon.common.presentation.model.response.ErrorResponse;
 import coz.weavon.common.presentation.model.response.RestResponse;
 import lombok.RequiredArgsConstructor;
