+++ src/main/java/coz/weavon/helper/MessageTranslator.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.application.service;
+package coz.weavon.helper;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.MessageSource;
